Moved BeanDefinition related method out of BeanFactory to AbstractBeanDefinitionRegistry
