+++ src/main/java/com/dimachine/core/AbstractBeanDefinitionRegistry.java
@@ -1,12 +1,13 @@
 package com.dimachine.core;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
-public abstract class AbstractBeanRegistry implements BeanRegistry {
-    protected final Map<BeanDefinition, Object> singletonBeans = new ConcurrentHashMap<>();
+public abstract class AbstractBeanDefinitionRegistry implements BeanDefinitionRegistry {
     protected final Set<BeanDefinition> beanDefinitions = Collections.newSetFromMap(new ConcurrentHashMap<>());
-    protected final DefaultObjectFactory objectFactory = new DefaultObjectFactory();
     protected final Set<String> beanNames = Collections.newSetFromMap(new ConcurrentHashMap<>());
 
     @Override
@@ -24,5 +25,24 @@ public abstract class AbstractBeanRegistry implements BeanRegistry {
                         " is already present in BeanFactory. Bean names should be unique.");
             }
         }
+    }
+
+    @Override
+    public <T> boolean containsBeanDefinitionOfType(Class<T> clazz) {
+        return beanDefinitions.stream()
+                .anyMatch(beanDefinition -> clazz.isAssignableFrom(beanDefinition.getBeanAssignableClass()));
+    }
+
+    @Override
+    public BeanDefinition getBeanDefinition(Class<?> beanClass) {
+        return beanDefinitions.stream()
+                .filter(beanDefinition -> beanClass.isAssignableFrom(beanDefinition.getBeanAssignableClass()))
+                .findFirst()
+                .orElseThrow(() -> new NoSuchBeanDefinitionException("No bean definition of type " + beanClass + " found"));
+    }
+
+    @Override
+    public Set<BeanDefinition> getBeanDefinitions() {
+        return Set.copyOf(beanDefinitions);
     }
 }
