+++ src/main/java/com/dimachine/core/BeanDefinitionMaker.java
@@ -1,5 +1,5 @@
 package com.dimachine.core;
 
-public interface BeanDefinitionMaker {
+public sealed interface BeanDefinitionMaker permits DefaultBeanDefinitionMaker {
     BeanDefinition makeBeanDefinition(String className);
 }
