+++ src/main/java/com/dimachine/core/BeanFactory.java
@@ -2,7 +2,6 @@ package com.dimachine.core;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 public interface BeanFactory extends AutoCloseable {
 
@@ -18,15 +17,11 @@ public interface BeanFactory extends AutoCloseable {
 
     <T> boolean contains(Class<T> clazz);
 
-    <T> boolean containsBeanDefinitionOfType(Class<T> clazz);
-
-    BeanDefinition getBeanDefinition(Class<?> beanClass);
-
     void registerSingleton(BeanDefinition beanDefinition, Object instance);
 
     <T> List<T> getAllBeansOfType(Class<T> clazz);
 
     <T> Map<String, T> getBeansMapOfType(Class<T> clazz);
 
-    Set<BeanDefinition> getBeanDefinitions();
+    void refresh();
 }
